+++ Middlewares/WebSocketMiddleware.cs
@@ -23,7 +23,7 @@ public class WebSocketMiddleware(RequestDelegate next) {
             if (context.WebSockets.IsWebSocketRequest) {
                 context.Request.EnableBuffering();
 
-                ushort? roomNumber = ushort.TryParse(context.Request.Query["roomNumber"].FirstOrDefault(), out var _rn)
+                uint? roomNumber = uint.TryParse(context.Request.Query["roomNumber"].FirstOrDefault(), out var _rn)
                     ? _rn : null;
 
                 var acc = Utilities.GetLoggedAccountFromContextOrNull();
